Refactor zoo components and localize limit warnings
